# libs/log: use fmt.Fprintf directly with *bytes.Buffer to avoid unnecessary allocations